Improve JFrog Pipelines templates (#325)
